requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
beautifulsoup4==4.14.2
billiard==4.2.3
boto3==1.29.7
botocore==1.32.7
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
Django==4.2.7
django-admin-rangefilter==0.13.3
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==4.3.14
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
kombu==5.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
Pillow==10.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
s3transfer==0.8.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.14
setuptools>=65.0.0
psycopg[binary]>=3.1.0